Complete Chap 1, Program 1.5, Problem 1 (page 12)
Solve the problem stated:
Write a java function `int maxargs(Stm s)` that tell the maximum number
of arguments of any `print` statement within any subexpression of a
fiven statement. For example, `maxargs(prog)` is 2.

Side Note: The style of the supplied classes set my teeth on edge. Also,
I hate having to use instanceof. It's so sloppy, not to mention slow.